Add BGP producer based on BGPKIT Broker
The BGPKIT broker and parser are an alternative to BGPStream.
To maintain compatibility with existing Kafka messages, both the input
parameters and output format of the messages is modeled after BGPStream
messages.
Not all fields are present for BGPKIT and the parser does not provide
records, but all that is relevant for AS Hegemony is included.